Consolidate time tricks into util/time_tricks
